Add Mozilla's CA Cert support to curl
Previously, curl wouldn't verify SSL certs when making outbound
connections.  It meant that it would either refuse to connect, or the
"-k" option needed to be added to acknowledge the insecurity and blindly
trust it. :/

This change adds the Mozilla CA store, repacked by upstream curl.  There
is also an update.sh script to update the pack if needed.

As this is Mozilla's store simply repacked, it continues to be under the
MPL 2.0 license.

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>