Add ACME Let's Encrypt HTTP path /.well-known/
Let's Encrypt offers free TLS/SSL certs, but to verify ownership of the
domain it is required to expose a path that their servers can reach us.

This is /.well-known/.  However, we don't want to have auth' on this
path (nor upgrade to https if the request was plain http).

All other requests remain unaffected, and even these paths still have
directory listing disabled.

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>